refactor: move loadConfiguration function
